combine queries tests
